improve: Optimize Pug mixins compiling
Prior to this commit, Pug mixin files were included in the
*src/pug/templates/_project_mixins.pug* file, which in turn was included
in the *src/pug/layout/_layout.pug* file. Therefore, any change in the
mixin file caused a rebuild of *_layout.pug*, and, consequently, all
pages of the project.

To avoid this problem, from now on the project mixin files will be
included directly where they are used.

It was decided to leave the inclusion of global mixin files in
*_layout.pug*, because these mixins do not change often.

The situation is similar with project and global variables.